pyproject.toml
[project]
name = "mcp-market-research"
version = "0.1.0"
description = "MCP server that generates market research questionnaires for Microsoft 365 Copilot Enterprise agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Market Research MCP Authors" }]
keywords = ["mcp", "market-research", "questionnaire", "copilot", "fastmcp"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "fastmcp>=2.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "openpyxl>=3.1",
  "pdfplumber>=0.11",
  "python-docx>=1.1",
  "docxtpl>=0.18",
  "deep-translator>=1.11",
  "babel>=2.15",
  "structlog>=24.1",
  "tenacity>=9.0",
  "httpx>=0.27",
  "starlette>=0.37",
  "reportlab>=4.2",
]

[project.optional-dependencies]
otel = [
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25",
  "opentelemetry-instrumentation-asgi>=0.46",
]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-openpyxl",
]

[project.scripts]
mcp-market-research = "mcp_market_research.__main__:main"

[project.urls]
Documentation = "https://github.com/example/mcp-market-research#readme"
Issues = "https://github.com/example/mcp-market-research/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_market_research"]

[tool.hatch.build.targets.wheel.force-include]
"src/mcp_market_research/samples" = "mcp_market_research/samples"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "PL", "RUF"]
ignore = ["PLR0913", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pdfplumber.*", "docxtpl.*", "deep_translator.*", "babel.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
markers = [
  "unit: fast unit tests",
  "integration: end-to-end tests touching the filesystem",
  "contract: MCP protocol contract tests",
]
